% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACuee.R
\name{MA_Cuee}
\alias{MA_Cuee}
\title{Online updating CUEE model averaging method}
\usage{
MA_Cuee(par_ori, data_train, data_test, p, e, n_block)
}
\arguments{
\item{par_ori}{The intitial parameter value}

\item{data_train}{The training dataset}

\item{data_test}{The test dataset}

\item{p}{The dimension of covarites used to construct the submodels}

\item{e}{The dimension of important covarites contained in every submodel}

\item{n_block}{The number of individuals at every stream data batch}
}
\value{
The CUEE parameter estimators of submodels, the CUEE model averaging weights and the predictive hazard ratio
}
\description{
The function provides the results of online updating CUEE model averaging method in stream datasets.
}
