% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAERenew.R
\name{MA_ERenew}
\alias{MA_ERenew}
\title{The extension of renewable model averaging method}
\usage{
MA_ERenew(par_ori, data_train, data_test, p, e, n_block)
}
\arguments{
\item{par_ori}{The intitial parameter value}

\item{data_train}{The training dataset}

\item{data_test}{The test dataset}

\item{p}{The dimension of covarites used to construct the submodels}

\item{e}{The dimension of important covarites contained in every submodel}

\item{n_block}{The number of individuals at every stream data batch}
}
\value{
The extended renewable parameter estimators of submodels, the extended renewable model averaging weights and the predictive hazard ratio
}
\description{
The function provides the results of the extended renewable model averaging method in stream datasets.
}
