% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefunctions.R
\name{packagedata}
\alias{packagedata}
\title{Read data}
\usage{
packagedata(file_name)
}
\arguments{
\item{file_name}{The file name of data}
}
\value{
The data of a example
}
\description{
The function that reads the data used in the example
}
